% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateIndexPedigree.R
\name{GenerateIndexPedigree}
\alias{GenerateIndexPedigree}
\title{GenerateIndexPedigree}
\usage{
GenerateIndexPedigree(
  age,
  sex,
  keep_future_children = TRUE,
  variant = "None",
  denovo_rate = 2e-06,
  is_denovo = NA,
  tell_family_prob = 0.7,
  tell_children_prob = 0.8,
  uptake_test_prob = 0.45,
  verbose = FALSE
)
}
\arguments{
\item{age}{numeric, age of index case}

\item{sex}{character, index case sex; either "M" or "F"}

\item{keep_future_children}{logical, whether to keep (TRUE) or remove (FALSE) children not yet born; default is TRUE}

\item{variant}{character, name of variant; default is "None" for unaffected index}

\item{denovo_rate}{numeric, between 0 and 1; rate of denovo mutation in given variant; default is 1/500000 or 2x10^(-6)}

\item{is_denovo}{logical, index mutation is de novo (TRUE) or inherited (FALSE); NA indicates denovo status is unknown and will be simulated; default is NA}

\item{tell_family_prob}{numeric, between 0 and 1; probability that an index case tells their family about their mutation; default is 0.70}

\item{tell_children_prob}{numeric, between 0 and 1; probability that an index case tells their children about their mutation, assumed this is <= tell_family_prob; default is 0.80}

\item{uptake_test_prob}{numeric, between 0 and 1; probability that a family member who is told about index mutation will get a genetic test; default is 0.45}

\item{verbose}{logical, whether to print messages during function; default is FALSE}
}
\value{
a data frame representing the family, where each row represents a person and columns represent attributes:

\describe{
\item{Age}{numeric, person age}
\item{Sex}{character, person sex; either "M" or "F"}
\item{RelationshipToIndex}{character, person's relationship to index}
\item{ChildOf}{character, which family member person is child of; NA for grandparents}
\item{DeNovo}{logical, NA/True whether mutation is de novo; only possible to be true for index and parent}
\item{Variant}{character, person's variant status; either NA or same as parameter}
\item{Told}{character, Yes/No, whether person was informed of index case}
\item{UptakeTesting}{character, Yes/No, on whether person chose to receive genetic health screening}
}
}
\description{
Function to simulate pedigree of index case to first cousins and back to grandparents; Genetic mutation of index is permeated through pedigree, and if the family members are told about mutation and if they uptake genetic testing is tracked
}
\details{
This function simulates pedigrees of individuals with a genetic mutation of interest, tracking family structure, variant permeation, and uptake of genetic testing. Family sizes are based on appropriate US census data, and due to correlation between number of children within pedigrees, this starts with assigning overall pedigree size to small, average, or large.

Children for each adult are generated using the SimulateChildren function.
}
\examples{
GenerateIndexPedigree(age = 35, variant = "BRCA1", sex = "M", is_denovo = NA)

}
\references{
Schweizer and Guzzo. Distributions of Age at First Birth, 1960-2018 (2020); https://www.bgsu.edu/ncfmr/resources/data/family-profiles/schweizer-guzzo-distribution-age-first-birth-fp-20-11.html

Khandwala, Zhang, Lu, and Eisenberg. The age of fathers in the USA is rising: an analysis of 168 867 480 births from 1972 to 2015, Human Reproduction, Volume 32, Issue 10, October 2017, Pages 2110–2116, https://doi.org/10.1093/humrep/dex267
}
\author{
Katelyn Queen \email{kqueen@mednet.ucla.edu}
}
