% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateChildren.R
\name{SimulateChildren}
\alias{SimulateChildren}
\title{SimulateChildren}
\usage{
SimulateChildren(
  mother_age,
  father_age = NA,
  pedigree_size = "Average",
  variant = "None",
  has_children = FALSE,
  known_child_age = NA,
  current_year = 2026
)
}
\arguments{
\item{mother_age}{numeric, age of mother}

\item{father_age}{numeric, age of father, if known; only required if simulating children based on father; default is NA}

\item{pedigree_size}{character, one of "Small", "Average", or "Large"; indicates average size of family in pedigree; default is "Average"}

\item{variant}{character, name of variant for index case; default is "None" for unaffected index}

\item{has_children}{logical, represents whether is known if parent has children; default is FALSE}

\item{known_child_age}{numeric, age of known child; only used if has_children is TRUE}

\item{current_year}{numeric, year to age individuals to; default is 2026}
}
\value{
a dataframe representing the children of indicated parent, where each row represents a person and columns represent attributes:

\describe{
\item{Age}{numeric, person age}
\item{Sex}{character, person sex; either "M" or "F"}
\item{Variant}{character, person's variant status; either NA or same as parameter}
\item{Father}{character, denotes which father child came from; accounts for complex family structures}
}
}
\description{
Function to simulate number and ages of children given age of mother or father
}
\details{
This function generates children given parental ages, pedigree size, variant status, and current year has_children and known_child_age are for generating past generations where we know there is at least one child of age X. If father_age given, calls .GenerateChildrenFromFather
}
\examples{
SimulateChildren(mother_age = 35, pedigree_size = "Small", variant = "BRCA1")

}
\references{
Guzzo KB. New Partners, More Kids: Multiple-Partner Fertility in the United States. Ann Am Acad Pol Soc Sci. 2014 Jul;654(1):66-86. doi: 10.1177/0002716214525571.

Monte and Knop, United States Census Bureau. Men's Fertility and Fatherhood: 2014 (2019); https://www.census.gov/content/dam/Census/library/publications/2019/demo/P70-162.pdf

Pew Research Center. The Modern American Family (2023); https://www.pewresearch.org/social-trends/2023/09/14/the-modern-american-family/

Schwartz, Doren, and Li, The Springer Series on Demographic Methods and Population Analysis, vol 51 (2020), DOI: 10.1007/978-3-030-48519-1_10

Schweizer. 30 Years of Change in Men's Entry into Fatherhood, 1987-2017 (2019); https://www.bgsu.edu/ncfmr/resources/data/family-profiles/schweizer-years-change-mens-entry-fatherhood-fp-19-28.html

Schweizer and Guzzo. Distributions of Age at First Birth, 1960-2018 (2020); https://www.bgsu.edu/ncfmr/resources/data/family-profiles/schweizer-guzzo-distribution-age-first-birth-fp-20-11.html

United States Census Bureau. Annual Social and Economic Supplements; https://www.census.gov/data/datasets/time-series/demo/cps/cps-asec.html
}
\author{
Katelyn Queen \email{kqueen@mednet.ucla.edu}
}
